+++ src/components/DatePicker/DatePicker.story.js
@@ -9,6 +9,7 @@ import { action } from '@storybook/addon-actions';
 import DatePicker from './DatePicker.js';
 import Input from '../Input';
 
+
 const stories = storiesOf('DatePicker', module);
 stories
   .addDecorator(withKnobs)